# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DreamCar Reserva de vehiculos",
        "description": "Esta API nos permitira ver y manipular informacion a las reservas de los vehiculos.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/rest/marcas/show/{id}": {
            "get": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Obtener una marcas por id",
                "description": "Renorna los detalles de una marca",
                "operationId": "0e93ee7d9ecd6de01f092ef16e78b11c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la marca",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la marca con id seleccionado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "marca": {
                                            "type": "string",
                                            "maxLength": 50,
                                            "example": "Nombre de una Marca",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rest/marcas/marca/{marca}": {
            "get": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Buscar marca por nombre",
                "description": "Retorna una marca",
                "operationId": "d14b980f7e53cd97cf48c06544735456",
                "parameters": [
                    {
                        "name": "marca",
                        "in": "path",
                        "description": "Nombre de la marca a buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marca encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "marca": {
                                                "type": "string",
                                                "maxLength": 50,
                                                "example": "Nombre de una Marca",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontro la marca",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}